+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 	Please see OpenShift release notes for official changes\n<!--Latest hash: 76cc46b739c27281db0f7b2a23eb091b2a70f698-->
 ## 4.14
 
+### Data Enhancement
+- [#741](https://github.com/openshift/insights-operator/pull/685) Gather openshift-monitoring persistent volumes
+
 ### Others
 - [#747](https://github.com/openshift/insights-operator/pull/747) Unit tests and file formatting for kube ctrl manager logs gatherer